Add Service annotations and reset semver
This PR resets the semver of the Flagsmith chart since its our own fork,
and henceforth will be dubbed `timescale-flagsmith`. More importantly,
we add the ability to interpolate annotations to both the
`flagsmith-frontend` and `flagsmith-api` Service objects. This will
allow us to properly annotate them with our external DNS annotations so
that Route 53 will generate a record.

In addition to the above, we also do the following:
- Update the main `README` to reflect our fork of this chart
- Add a script for generating Helm packages